feat(config): Add cloth config compat for (Neo)Forge
* Move fabric ModMenuIntegration to src/main/fabric/config

* Cloth Config hasn't been updated for Forge, so it isn't tested
