+++ forge/src/main/java/com/aizistral/nochatreports/forge/NoChatReports.java
@@ -1,27 +1,41 @@
 package com.aizistral.nochatreports.forge;
 
 import java.nio.file.Path;
+import java.util.function.Function;
 
+import com.aizistral.nochatreports.common.NCRClient;
 import com.aizistral.nochatreports.common.NCRCore;
 import com.aizistral.nochatreports.common.platform.PlatformProvider;
 import com.aizistral.nochatreports.common.platform.events.ClientEvents;
 
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.ConfigScreenHandler;
 import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
 import net.minecraftforge.eventbus.api.listener.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLEnvironment;
 import net.minecraftforge.fml.loading.FMLPaths;
 
 @Mod("nochatreports")
 public class NoChatReports implements PlatformProvider {
 	
-	public NoChatReports() {
+	public NoChatReports(FMLJavaModLoadingContext context) {
 		NCRCore.awaken(this);
+        Function<Screen, Screen> configScreenFactory = NCRClient.getConfigScreen();
+        if (isOnClient() && configScreenFactory != null) {
+            context.registerExtensionPoint(
+                ConfigScreenHandler.ConfigScreenFactory.class,
+                () -> new ConfigScreenHandler.ConfigScreenFactory(
+                        (minecraft, parent) -> configScreenFactory.apply(parent)
+                )
+            );
+        }
 	}
 	
 	@Override
