+++ neoforge/src/main/java/com/aizistral/nochatreports/neoforge/NoChatReports.java
@@ -1,25 +1,37 @@
 package com.aizistral.nochatreports.neoforge;
 
+import com.aizistral.nochatreports.common.NCRClient;
 import com.aizistral.nochatreports.common.NCRCore;
 import com.aizistral.nochatreports.common.platform.PlatformProvider;
 import com.aizistral.nochatreports.common.platform.events.ClientEvents;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.neoforged.api.distmarker.Dist;
+import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.neoforge.client.event.ClientPlayerNetworkEvent;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.loading.FMLEnvironment;
 import net.neoforged.fml.loading.FMLPaths;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 
 import java.nio.file.Path;
+import java.util.function.Function;
 
 @Mod("nochatreports")
 public class NoChatReports implements PlatformProvider {
 	
 	public NoChatReports() {
 		NCRCore.awaken(this);
+        Function<Screen, Screen> configScreenFactory = NCRClient.getConfigScreen();
+        if (isOnClient() && configScreenFactory != null) {
+            ModLoadingContext.get().registerExtensionPoint(
+                    IConfigScreenFactory.class,
+                    () -> (minecraft, parent) -> configScreenFactory.apply(parent)
+            );
+        }
 	}
 	
 	@Override
