+++ src/main/resources/fabric.mod.json
@@ -29,7 +29,7 @@
       "com.aizistral.nochatreports.fabric.NoChatReportsClient"
     ],
     "modmenu": [
-      "com.aizistral.nochatreports.common.config.ModMenuIntegration"
+      "com.aizistral.nochatreports.fabric.config.ModMenuIntegration"
     ]
   },
   "mixins": [
